Added a master file ingester script
Script that will "ingest" files based on SHA1 checksum into a single
directory tree, split by 4-character slices of the SHA1. This way only a
maximum of 16^4 files will ever be contained in a single directory,
which is managable. This comes at the cost of more directory creations,
so more total disk space will be used.